# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: ДомБыт API
  version: 1.0.0
  description: API для специализаций и заказов
servers:
  - url: /

tags:
  - name: specialization
    description: Управление специализациями
  - name: order
    description: Управление заказами

paths:
  /api/specializations:
    get:
      tags: [ specialization ]
      operationId: getAllSpecializations
      summary: Получить все специализации
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpecializationResponse"
    post:
      tags: [ specialization ]
      operationId: createSpecialization
      summary: Создать специализацию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpecializationRequest"
      responses:
        "201":
          description: Created

  /api/specializations/{id}:
    get:
      tags: [ specialization ]
      operationId: getSpecializationById
      summary: Получить специализацию по ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
    put:
      tags: [ specialization ]
      operationId: updateSpecialization
      summary: Обновить специализацию
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpecializationRequest"
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
    delete:
      tags: [ specialization ]
      operationId: deleteSpecialization
      summary: Удалить специализацию
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found

  /api/orders:
    get:
      tags: [ order ]
      operationId: getOrders
      summary: Получить заказы пользователя
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
    post:
      tags: [ order ]
      operationId: createOrder
      summary: Создать заказ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created

  /api/orders/{id}:
    get:
      tags: [ order ]
      operationId: getOrderById
      summary: Получить заказ по ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
    put:
      tags: [ order ]
      operationId: updateOrder
      summary: Обновить заказ
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: OK
    delete:
      tags: [ order ]
      operationId: cancelOrder
      summary: Отменить заказ
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: No Content

  /api/orders/{id}/status:
    patch:
      tags: [ order ]
      operationId: updateOrderStatus
      summary: Изменить статус заказа
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: OK

  /api/orders/available:
    get:
      tags: [ order ]
      operationId: getAvailableOrders
      summary: Доступные заказы для мастера
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"

components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    SpecializationResponse:
      type: object
      required: [ id, name, basePrice, weatherDependent ]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        basePrice:
          type: number
          format: double
        weatherDependent:
          type: boolean

    CreateSpecializationRequest:
      type: object
      required: [ name, basePrice ]
      properties:
        name:
          type: string
        description:
          type: string
        basePrice:
          type: number
        weatherDependent:
          type: boolean

    UpdateSpecializationRequest:
      type: object
      required: [ name, basePrice ]
      properties:
        name:
          type: string
        description:
          type: string
        basePrice:
          type: number
        weatherDependent:
          type: boolean

    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        address:
          type: string
        status:
          type: string
        price:
          type: number
        weatherCoefficient:
          type: number
        scheduledDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time

    CreateOrderRequest:
      type: object
      required: [ specializationId, title, address, scheduledDate ]
      properties:
        specializationId:
          type: integer
        title:
          type: string
        description:
          type: string
        address:
          type: string
        scheduledDate:
          type: string
          format: date-time

    UpdateOrderRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        address:
          type: string

    UpdateStatusRequest:
      type: object
      required: [ status ]
      properties:
        status:
          type: string
          enum: [ ASSIGNED, IN_PROGRESS, COMPLETED, CANCELLED ]